more pep8 and iter divertor docs moved
