+++ paramak/parametric_reactors/ball_reactor.py
@@ -1,8 +1,6 @@
 
 import warnings
 
-import cadquery as cq
-
 import paramak
