+++ paramak/parametric_reactors/single_null_submersion_reactor.py
@@ -1,6 +1,4 @@
 
-import cadquery as cq
-
 import paramak
