+++ paramak/parametric_shapes/extruded_circle_shape.py
@@ -1,6 +1,5 @@
-import math
+
 from collections import Iterable
-from hashlib import blake2b
 
 import cadquery as cq
 
@@ -12,20 +11,25 @@ class ExtrudeCircleShape(Shape):
 
     Args:
         points (tuple containing X (float), Z (float) value for the central point): a list
-            of a single XZ coordinate which is the central point of the circle. For example
-            [(10, 10)]
+            of a single XZ coordinate which is the central point of the circle.
+            For example [(10, 10)]
         radius (float): the radius of the circle
-        stp_filename (str): the filename used when saving stp files as part of a reactor
+        stp_filename (str): the filename used when saving stp files as part of a
+            reactor
         color (RGB or RGBA - sequences of 3 or 4 floats, respectively, each in the range 0-1):
             the color to use when exporting as html graphs or png images
-        distance (float): the extrusion distance to use (cm units if used for neutronics)
-        azimuth_placement_angle (float or iterable of floats): the angle or angles to use when
-            rotating the shape on the azimuthal axis
-        cut (CadQuery object): an optional CadQuery object to perform a boolean cut with
-            this object
-        material_tag (str): the material name to use when exporting the neutronics descrption
-        name (str): the legend name used when exporting a html graph of the shape
-        workplane (str): the orientation of the CadQuery workplane. Options are XY, YZ, XZ.
+        distance (float): the extrusion distance to use (cm units if used for
+            neutronics)
+        azimuth_placement_angle (float or iterable of floats): the angle or
+            angles to use when rotating the shape on the azimuthal axis
+        cut (CadQuery object): an optional CadQuery object to perform a boolean
+            cut with this object
+        material_tag (str): the material name to use when exporting the
+            neutronics descrption
+        name (str): the legend name used when exporting a html graph of the
+            shape
+        workplane (str): the orientation of the CadQuery workplane. Options are
+            XY, YZ, XZ.
 
     Returns:
         a paramak shape object: a Shape object that has generic functionality
