+++ paramak/parametric_shapes/extruded_mixed_shape.py
@@ -1,5 +1,4 @@
 from collections import Iterable
-from hashlib import blake2b
 
 import cadquery as cq
