+++ paramak/parametric_shapes/extruded_straight_shape.py
@@ -1,6 +1,5 @@
 
 from collections import Iterable
-from hashlib import blake2b
 
 import cadquery as cq
