+++ tests/test_SubmersionTokamak.py
@@ -3,8 +3,6 @@ import unittest
 from pathlib import Path
 import warnings
 
-import pytest
-
 import paramak
 
 
@@ -337,7 +335,7 @@ class test_SubmersionTokamak(unittest.TestCase):
 
         def warning_trigger():
             try:
-                test_reactor = paramak.SubmersionTokamak(
+                paramak.SubmersionTokamak(
                     inner_bore_radial_thickness=25,
                     inboard_tf_leg_radial_thickness=50,
                     center_column_shield_radial_thickness=50,
